[TASK] Expand short-hand links
Implements a search-replace for shorthand links to only a file's file name.

Since each step-by-step guide should have a unique title, it follows that a file name based on the title should also be unique. This allows us to link to a file by only its name, and we don't have to think about the folder structure.